Rename test dir to tests and move unit and integration tests
Refactoring

Revert and refactor unit tests

Finish moving tests to separate modules

Remove test feature files

Change paths for tests

Change makefile

Finish moving tests to separate modules

Remove test feature files
