+++ tests/steps/application_v2_steps.py
@@ -10,7 +10,7 @@ from algosdk.abi.contract import NetworkInfo
 from algosdk.error import ABITypeError, IndexerHTTPError
 from algosdk.future import transaction
 
-from test.steps.other_v2_steps import read_program
+from tests.steps.other_v2_steps import read_program
 
 
 def operation_string_to_enum(operation):
